fix: place의 좌표값 double로 타입 변경

+++ plogging-service/src/main/java/com/songspasssta/ploggingservice/dto/response/PlaceResponse.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public record PlaceResponse(@JsonProperty("searchedPlace") List<PlaceDto> searchedPlace) {
 
-    public record PlaceDto(Float latitude, Float longitude, String roadAddr, String placeInfo) {
+    public record PlaceDto(Double latitude, Double longitude, String roadAddr, String placeInfo) {
     }
 }
